third_party/WebKit/LayoutTests/fullscreen/api/document-exit-fullscreen-vs-request.html

<!DOCTYPE html>
<title>Document#exitFullscreen() vs. Element#requestFullscreen()</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<div id="parent"><div><div></div></div></div>
<script>
// Adapted from https://github.com/w3c/web-platform-tests/pull/4250
// TODO(foolip): Remove this test when the above is imported and passing.
async_test(t => {
  const parent = document.getElementById("parent");
  const child = parent.firstChild;
  const grandChild = child.firstChild;

  document.onfullscreenchange = t.step_func(() => {
    // We are now in fullscreen, so exiting requires a resize but requesting
    // does not.
    assert_equals(document.fullscreenElement, parent, "fullscreenElement after fullscreenchange event");

    trusted_click(t.step_func(() => {
      // Request fullscreen on another element, to avoid any synchronous
      // short-circuiting on document.fullscreenElement.requestFullscreen(),
      // which used to be in the spec. Also request both before and after the
      // exit. Both requests synchronously enqueue animation frame tasks. They
      // may run after exiting, but still before the animation frame task for
      // the exit, and so both will succeed, and there will be 3
      // fullscreenchange events, but not matching the order of the calls.

      let i = 0;
      const expected = [child, grandChild, null];
      document.onfullscreenchange = t.step_func(() => {
        assert_equals(document.fullscreenElement, expected[i], "fullscreenElement when i=" + i);
        i++;
        if (i == 3)
          t.done();
      });

      child.requestFullscreen();
      assert_equals(document.fullscreenElement, parent, "fullscreenElement after first requestFullscreen()");
      document.exitFullscreen();
      assert_equals(document.fullscreenElement, parent, "fullscreenElement after exitFullscreen()");
      grandChild.requestFullscreen();
      assert_equals(document.fullscreenElement, parent, "fullscreenElement after second requestFullscreen()");
    }), parent);
  });
  document.onfullscreenerror = t.unreached_func("fullscreenerror event");

  trusted_request(parent);
});
</script>
